Implement `clearAppState` in `XCTestIOSDevice`
