+++ maestro-ios/src/main/java/ios/xcrun/Simctl.kt
@@ -106,4 +106,60 @@ object Simctl {
 
         reboot(deviceId)
     }
+
+    fun terminate(deviceId: String, bundleId: String) {
+        // Ignore error return: terminate will fail if the app is not running
+        ProcessBuilder(
+            listOf(
+                "xcrun",
+                "simctl",
+                "terminate",
+                deviceId,
+                bundleId
+            )
+        )
+            .start()
+            .waitFor()
+    }
+
+    fun clearAppState(deviceId: String, bundleId: String) {
+        // Stop the app before clearing the file system
+        // This prevents the app from saving its state after it has been cleared
+        terminate(deviceId, bundleId)
+
+        // Wait for the app to be stopped
+        Thread.sleep(1500)
+
+        // deletes app data, including container folder
+        val appDataDirectory = getApplicationDataDirectory(deviceId, bundleId)
+        ProcessBuilder(listOf("rm", "-rf", appDataDirectory)).start().waitFor()
+
+        // forces app container folder to be re-created
+        val paths = listOf(
+            "Documents",
+            "Library",
+            "Library/Caches",
+            "Library/Preferences",
+            "SystemData",
+            "tmp"
+        )
+
+        val command = listOf("mkdir", appDataDirectory) + paths.map { "$appDataDirectory/$it" }
+        ProcessBuilder(command).start().waitFor()
+    }
+
+    private fun getApplicationDataDirectory(deviceId: String, bundleId: String): String {
+        val process = ProcessBuilder(
+            listOf(
+                "xcrun",
+                "simctl",
+                "get_app_container",
+                deviceId,
+                bundleId,
+                "data"
+            )
+        ).start()
+
+        return String(process.inputStream.readBytes()).trimEnd()
+    }
 }
