+++ maestro-ios/src/main/java/ios/xctest/XCTestIOSDevice.kt
@@ -4,13 +4,18 @@ import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
+import com.github.michaelbull.result.map
 import com.github.michaelbull.result.runCatching
 import hierarchy.Error
 import hierarchy.XCUIElement
 import idb.Idb
+import idb.fileContainer
+import idb.mkdirRequest
+import idb.rmRequest
 import ios.IOSDevice
 import ios.IOSScreenRecording
 import ios.device.DeviceInfo
+import ios.xcrun.Simctl
 import maestro.logger.Logger
 import okio.Sink
 import okio.buffer
@@ -157,7 +162,8 @@ class XCTestIOSDevice(
     }
 
     override fun clearAppState(id: String): Result<Unit, Throwable> {
-        error("Not supported")
+        deviceId?.let { Simctl.clearAppState(deviceId, id) }
+        return Ok(Unit)
     }
 
     override fun clearKeychain(): Result<Unit, Throwable> {
@@ -169,7 +175,8 @@ class XCTestIOSDevice(
     }
 
     override fun stop(id: String): Result<Unit, Throwable> {
-        error("Not supported")
+        deviceId?.let { deviceId -> Simctl.terminate(deviceId, id) }
+        return Ok(Unit)
     }
 
     override fun openLink(link: String): Result<Unit, Throwable> {
